Added support for QCAD's CXF format font files.  It's pretty basic support at the moment.  To use it, one must first select the font from the main properties page -> Font Settings -> Active Font.
When text is added (in the usual way), the chosen font is used.  The text that is placed can then be used (via a right mouse button) to 'Make a Sketch'.

I don't propose to add the font files for delivery with HeeksCAD as they're covered by GPL but this code is completely original.  If the QCAD fonts are delivered separately to HeeksCAD then this code can read and use it.  On Ubuntu Linux the fonts are delivered with the qcad-data product.  I don't know how it can be delivered for Windows.  In fact it doesn't even have a place to look in the Windows environment so it will simply fail to load them.  The function then just reverts to the previous font functionality.  i.e. OpenGL font.  If I'm making a mistake on the licensing stuff here, please let me know and I will remove this code.  I think it's alright as the only code I have added to the project is completely original.  I don't think the fact that this code reads files that are covered by GPL is an issue.  Again, if someone disagrees then please let me know.

My hope with this is to be able to place text in a drawing, convert it to Sketches and then use them with profiles to engrave text into a workpiece.  It's not there yet for all this.  I think I'm embedding sketches within sketches within sketches at the moment.

It's a start in any case.