Added a link between the geometry tolerance value and the number of decimal places used to write double values in XML.  By default the %lf format uses 6 decimal places while OpenCascade has a default tolerance that includes the seventh decimal place.  The number of decimal places used in XML will now be one more than is required by the configured geometry tolerance.
Changed the Sketch to Wire conversion routine to use Vertex objects rather than gp_Pnt objects so that the geometry tolerance could be included in the edge object's definition.  This ensures that the geometry tolerance used in the XML data is consistent with that used to detect edge connections in the OpenCascade libraries.