Fixed problem where the characters were not being rotated about the normal of the sketch correctly.
This function is getting closer for release but it's still not quite
right.  If the original text string has its location at the origin then
the translated version works fine.  If, however, the text is
scaled and/or moved then the rendered version makes less sense.

It mostly works for aligning text with sketches.  eg: for drawing
text around a circle or along a some other sketch.  It has been
written as a general orientation modifier class.  Eventually, I
hope to allow arrays of sketch objects to be aligned and/or
rotated along another sketch's path.  A modification of
the alignment sketch, re-orients the array of objects.  That's the idea
anyway.