New branch pycad with new CMakeLists file
danielfalck suggested that I copy the existing CMakeLists.txt file
in src/ to PyHeeksCAD/ and edit it appropriately to include Dan's
pythonheekscad code. So I've made a start on this. But I just
couldn't stand the odd four-column layout and, since it's a new file,
there are no whitespace arguments. I went single-column so I can
open the file twice in two vertical windows and do a side-by-side
compare, I found that to be a lot easier to do and easier to
find errors. No doubt there are still plenty of errors though.


Signed-off-by: Kim Kirwan <Kim@KimKirwan.com>